Payment API Done

+++ IOSApp/cmpe202App/ViewControllers/SearchViewController.swift
@@ -147,7 +147,7 @@ class SearchViewController: UIViewController,UITableViewDelegate,UITableViewData
 
             cell.hotelName.text = dataDict["name"] as? String
             cell.location.text = dataDict["city"] as? String
-            let price = dataDict["basePrice"] as! Int
+            let price = dataDict["basePrice"] as! Float
             cell.price.text = "from $\(String(describing: price))"
             let url = URL(string: dataDict["imageURL"] as! String)!
             cell.hotelImage.af.setImage(withURL: url, cacheKey: "searchTable\(indexPath.row)", placeholderImage: UIImage (named: "tableListImage"), serializer: nil, filter: nil, progress:nil, progressQueue: .global(), imageTransition: .noTransition, runImageTransitionIfCached: false, completion: nil)
